make data type bug when calling std::min
in some i686 platform, sizt_t is defined as unsign int 
instead of unsigned long. therefore passing one parameter as 
integer suffix 2ul or 3ul and the other parameter as sizt_t 
can not compile in some i686 platforms

+++ OptionParser.cpp
@@ -355,7 +355,7 @@ OptionParser::help_message() const {
   }
   
   ss << endl << "Help options:" << endl;
-  for (size_t i = 0; i < std::min(2ul, options.size()); ++i)
+  for (size_t i = 0; i < std::min(static_cast<size_t>(2), options.size()); ++i)
     ss << SPACE_BEFORE_SHORT << std::left << std::setw(max_name_len) 
        << option_names[i] << SPACE_BTWN_SHRT_LNG
        << options[i].format_option_description(max_name_len +
